Added Sign Owner Whitelist
I have create an AllowedIDs.json file that you can add SteamIDs to.
These SteamIDs are the Owners' of the signs that you want to teleport people. This is to make it so not everyone can just place a sign and teleport.

This also adds a new Permission "signteleporter.teleport", meaning only players with this Permission can be teleported to a sign.